requirements.txt
asgiref==3.2.3
autopep8==1.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==3.0.3
django-braces==1.14.0
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.9
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.1
urllib3==1.25.8